Refactor

+++ app/Http/Controllers/KelTaniController.php
@@ -3,12 +3,10 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
-use Crypter;
-
 use App\KelompokTani;
 use App\Komoditas;
 use App\Penyuluh;
+use HCrypt;
 
 class KelTaniController extends Controller
 {
@@ -17,44 +15,47 @@ class KelTaniController extends Controller
     return view('KelompokTani.Data', compact('KelompokTani'));
   }
 
-  public function Tambah(){
+  public function TambahForm(){
     $Komoditas = Komoditas::all();
     $Penyuluh = Penyuluh::all();
     return view('KelompokTani.Tambah', compact('Komoditas', 'Penyuluh'));
   }
 
-  public function submitTambah(Request $request){
+  public function TambahSubmit(Request $request){
     $KelompokTani = new KelompokTani;
     $KelompokTani->fill($request->all());
     $KelompokTani->save();
     foreach ($request->komoditas_id as $KomoditasId) {
       $KelompokTani->Komoditas()->attach($KomoditasId);
     }
-    return redirect()->Route('Data-Kelompok-Tani')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Tambah Data Berhasil']);
+    return redirect()->Route('kelompokTaniData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Tambah Data Berhasil']);
   }
 
-  public function Edit($Id){
-    $Id = Crypter::Decrypt($Id);
+  public function EditForm($Id){
+    $Id = HCrypt::Decrypt($Id);
     $KelompokTani = KelompokTani::findOrFail($Id);
     $Komoditas = Komoditas::all();
     $Penyuluh = Penyuluh::all();
     return view('KelompokTani.Edit', compact('KelompokTani', 'Komoditas', 'Penyuluh'));
   }
 
-  public function submitEdit(Request $request, $Id){
-    $Id = Crypter::Decrypt($Id);
+  public function EditSubmit(Request $request, $Id){
+    $Id = HCrypt::Decrypt($Id);
     $KelompokTani = KelompokTani::findOrFail($Id);
     $KelompokTani->fill($request->all());
     $KelompokTani->Komoditas()->sync($request->komoditas_id);
     $KelompokTani->save();
-    return redirect()->Route('Data-Kelompok-Tani')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
+    return redirect()->Route('kelompokTaniData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
   }
 
-  public function Delete($Id){
-    $Id = Crypter::Decrypt($Id);
-    $KelompokTani = KelompokTani::findOrFail($Id);
-    $KelompokTani->Komoditas()->detach();
-    $KelompokTani->delete();
-    return redirect()->Route('Data-Kelompok-Tani')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Hapus Data Berhasil']);
+  public function Hapus($Id=null,$Verify=null){
+    if ($Verify) {
+      $Id = HCrypt::Decrypt($Id);
+      $KelompokTani = KelompokTani::findOrFail($Id);
+      $KelompokTani->Komoditas()->detach();
+      $KelompokTani->delete();
+      return redirect()->Route('kelompokTaniData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Hapus Data Berhasil']);
+    }
+    return abort(404);
   }
 }
