+++ app/Http/Controllers/UnitKerjaController.php
@@ -3,9 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Crypter;
-
 use App\UnitKerja;
+use HCrypt;
 
 class UnitKerjaController extends Controller
 {
@@ -14,35 +13,38 @@ class UnitKerjaController extends Controller
     return view('UnitKerja.Data', compact('UnitKerja'));
   }
 
-  public function Tambah(){
+  public function TambahForm(){
     return view('UnitKerja.Tambah');
   }
 
-  public function submitTambah(Request $request){
+  public function TambahSubmit(Request $request){
     $UnitKerja = new UnitKerja;
     $UnitKerja->fill($request->all());
     $UnitKerja->save();
-    return redirect()->Route('Data-Unit-Kerja')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Tambah Data Berhasil']);
+    return redirect()->Route('unitKerjaData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Tambah Data Berhasil']);
   }
 
-  public function Edit($Id){
-    $Id = Crypter::Decrypt($Id);
+  public function EditForm($Id){
+    $Id = HCrypt::Decrypt($Id);
     $UnitKerja = UnitKerja::findOrFail($Id);
     return view('UnitKerja.Edit', compact('UnitKerja'));
   }
 
-  public function submitEdit(Request $request, $Id){
-    $Id = Crypter::Decrypt($Id);
+  public function EditSubmit(Request $request, $Id){
+    $Id = HCrypt::Decrypt($Id);
     $UnitKerja = UnitKerja::findOrFail($Id);
     $UnitKerja->fill($request->all());
     $UnitKerja->save();
-    return redirect()->Route('Data-Unit-Kerja')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
+    return redirect()->Route('unitKerjaData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Edit Data Berhasil']);
   }
 
-  public function Delete($Id){
-    $Id = Crypter::Decrypt($Id);
-    $UnitKerja = UnitKerja::findOrFail($Id);
-    $UnitKerja->delete();
-    return redirect()->Route('Data-Unit-Kerja')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Delete Data Berhasil']);
+  public function Hapus($Id=null,$Verify=null){
+    if ($Verify) {
+      $Id = HCrypt::Decrypt($Id);
+      $UnitKerja = UnitKerja::findOrFail($Id);
+      $UnitKerja->delete();
+      return redirect()->Route('unitKerjaData')->with(['alert' => true, 'tipe' => 'success', 'judul' => 'Berhasil', 'pesan' => 'Delete Data Berhasil']);
+    }
+    return abort(404);
   }
 }
